+++ examples/laravel/app/Telegram/Commands/StartCommand.php
@@ -15,6 +15,7 @@ readonly class StartCommand implements TelegramWebhookCommandHandler
         //
     }
 
+    #[\Override]
     public function handle(TelegramWebhookCommand $command, TelegramWebhookUpdate $update, string $botName): mixed
     {
         $chatId = $command->message()->chat()?->id();